feat(TD-8): Add backend logic to create and get tasks
