+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
-      resources :tasks, only: [:index, :show, :create, :update, :destroy]
+      resources :tasks, only: [:index, :show, :create]
     end
   end
 end
